requirements.txt
configparser==5.0.0
RPi.bme280==0.2.3
smbus2==0.3.0
virtualenv==16.7.9
